+++ CHANGELOG.md
@@ -1,14 +1,14 @@
 # Changelog
 
 # v2.0.0
-Added modes: `speed` or `precision`. Speed by default.
+Added modes: `speed` and `precision`. Speed by default.
 ```
 const fac = new FastAverageColor(image, {mode: 'precision'});
 ```
 Added examples:
 - [Box shadow, 4 sides](https://hcodes.github.io/fast-average-color/examples/box-shadow-4-sides.html)
 - [Gradient](https://hcodes.github.io/fast-average-color/examples/gradient.html)
-- [Video](https://hcodes.github.io/fast-average-color/examples/video.html)
+- [Ambilight](https://hcodes.github.io/fast-average-color/examples/ambilight.html)
 - [Define the average color for your images](https://hcodes.github.io/fast-average-color/examples/define.html)
 
 Added support for HTMLVideoElement.